requirements.txt
certifi==2018.11.29
Pillow==5.3.0
